// Client/projects/coi/src/app/configuration/form-builder-create/form-builder-create-interface.ts
export class FormIntegration {
    formModuleCode = null;
    formSubSectionCode = null;
    formRule = null;
    formUsageId = null;
    formOrderNumber = null;
}



export interface GetAllFormUsage {
    formUsageId: number
    formBuilderId: number
    formOrderNumber: number
    moduleCode: string
    subModuleCode: string
    businessRuleId: any
    description: string
    isActive: string
    createTimestamp: string
    createUser: string
    updateTimestamp: string
}

export interface SubModules {
    SUB_MODULE_CODE: number
    DESCRIPTION: string
    MODULE_CODE: number
    IS_ACTIVE: string
  }

  export interface FetchModuleCode {
    applicableQuestionnaire: any
    questionnaireId: any
    moduleItemKey: any
    moduleSubItemKey: any
    moduleItemCode: any
    moduleSubItemCode: any
    questionnaireAnswerHeaderId: any
    questionnaireAnsAttachmentId: any
    questionnaireCompleteFlag: any
    actionUserId: any
    actionPersonId: any
    actionPersonName: any
    acType: any
    questionnaireName: any
    newQuestionnaireVersion: boolean
    questionEditted: boolean
    questionnaireList: any
    questionnaireGroup: any
    header: any
    questionnaire: any
    usage: any
    fileName: any
    fileContent: any
    length: any
    remaining: any
    fileTimestamp: any
    contentType: any
    personId: any
    multipartFile: any
    moduleList: ModuleList[]
    isInserted: any
    updateTimestamp: any
    copyModuleItemKey: any
    questionnaireNumbers: any
    lookUpDetails: any
    newQuestionnaireId: any
    moduleSubItemCodes: any[]
    questionnaireBusinessRules: any
    ruleId: any
    rulePassed: any
    questionnaireMode: any
  }

  export interface ModuleList {
    SUB_MODULE_CODE: number
    DESCRIPTION: string
    MODULE_CODE: any
    IS_ACTIVE: string
  }

  export class saveFormUsage {
    formBuilderNumber: string
    formBuilderId: string
    moduleCode: any
    subModuleCode: any
    businessRuleId: any
    description: string
    isActive: string
  }

  export interface IntegrtionResponse {
    formUsageId: number
    formBuilderId: number
    formOrderNumber: number
    moduleCode: string
    subModuleCode: string
    businessRuleId: any
    description: string
    isActive: string
    createTimestamp: string
    createUser: string
    updateTimestamp: string
    updateUser: string
  }

  export class UpdateFormUsage {
    formUsageId: any;
    formBuilderNumber: string;
    formBuilderId: string;
    moduleCode: any;
    subModuleCode: any;
    businessRuleId: any;
    description: string;
    isActive: string;
    formOrderNumber: any;
  }